LPD-63103 add performance test for CTSQLTransformerImpl
